Change default template actions to be buttons
A lot of Bootstrap varied themes will change a, a:hover, and a:visited
to match the theme. Which always throws off the text in the .btn colors
of the modal. The user of backbone.bootstrap-modal can provide a custom
defined template but this increases code bloat where it doesn't seem
necessary.